+++ packages/core/src/scuttle.d.ts
@@ -2,8 +2,11 @@
  * Options for scuttling global properties
  */
 export type LavaMoatScuttleOpts = {
+  /** Whether scuttling is enabled or not. */
   enabled: boolean;
+  /** List of properties to exclude from scuttling. */
   exceptions?: Array<string | RegExp>;
+  /** Name of the scuttler function to use which is expected to be found as a property on the global object (e.g. if scuttlerName is 'x', scuttler function is obtained from globalThis['x']). */
   scuttlerName?: string;
 }