+++ packages/core/src/types.ts
@@ -6,6 +6,6 @@ export * from './index'
 export type * from './moduleRecord'
 export type * from './options'
 export type * from './parseForPolicy'
-export type * from './scuttle'
+export type { LavaMoatScuttleOpts } from './scuttle.d.ts'
 export { E as EndowmentsToolkitFactory }